Allow empty string text nodes
Currently, if you try to render a node like this:
```js
text('')
```
It will go through this part of the code to get HTML escaped, however, because the `||` operator considers `''` a falsy value, it goes for `node.name` instead, which does not exist on a VDOM node. The first line in this `escapeHtml` function converts this `undefined` value to string, which ends up being the string `'undefined'`.

I think this is a bug?

Hope this helps!